Class: AddUserToRoleRequest, AddUserToRoleRespone
