+++ TaskManager.Data/Context/ApplicationDbContext.cs
@@ -1,5 +1,7 @@
 ﻿using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Storage;
 using TaskManager.Models.Entities;
 using Task = TaskManager.Models.Entities.Task;
 
@@ -10,7 +12,26 @@ namespace TaskManager.Data.Context
         public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options)
            : base(options)
         {
+            try
+            {
+                var databasecreater = Database.GetService<IDatabaseCreator>() as RelationalDatabaseCreator;
+                if (databasecreater != null)
+                {
+                    if (!databasecreater.CanConnect()){
+                        databasecreater.Create();
+                    }
+                    if (!databasecreater.HasTables())
+                    {
+                        databasecreater.CreateTables();
+                    }
+                }
+            }catch (Exception ex)
+            {
+                throw new Exception(ex.Message);
+            }
         }
+
+
         public DbSet<Project> Projects { get; set; }
         public DbSet<Task> Tasks { get; set; }
         public DbSet<Notification> Notifications { get; set; }
