+++ TaskManager.Services/Interfaces/IRoleService.cs
@@ -1,5 +1,7 @@
 ﻿using TaskManager.Services.Infrastructure;
 using TaskManager.Services.Implementations;
+using TaskManager.Models.Dtos.Request;
+using TaskManager.Models.Dtos.Response;
 
 namespace TaskManager.Services.Interfaces
 {
