+++ TaskManagerApi/appsettings.json
@@ -9,7 +9,7 @@
   "AllowedHosts": "*",
 
   "ConnectionStrings": {
-    "DefaultConnection": "Data Source=JERMAINE\\SQLEXPRESS; Initial Catalog=TaskManagerDb;Integrated Security=True;Connect Timeout=30;Encrypt=False;TrustServerCertificate=False;ApplicationIntent=ReadWrite;MultiSubnetFailover=False"
+    "DefaultConnection": "Data Source=JERMAINE\\SQLEXPRESS;Initial Catalog=TaskManagerDb;Password= #Alexissanchez31; Integrated Security=True;Connect Timeout=30;Encrypt=False;TrustServerCertificate=True;ApplicationIntent=ReadWrite;MultiSubnetFailover=False"
   },
 
   "JwtConfig": {
